feat(collections): created tests
feat(collection-colums): created collection column tests
feat(collection-table): created tests for collection table methods
feat(collection-table-column): create tests for collection table column repository
feat(collection-item): created tests for collection item repository
feat(string-extensions): created tests for custom string extensions
feat(collectionservice): added tests for collectionservice
feat(collection-column-service): created tests for collectionColumnService
feat(collection-item-service): created tests for collectionItemService
feat(column-handler-strategy): created tests for column handler strategy
feat(jacoco): configured test coverage
fix(tests): fixed tests and code on manual testing

+++ src/main/kotlin/io/webcontify/backend/collections/services/CollectionItemService.kt
@@ -1,5 +1,6 @@
 package io.webcontify.backend.collections.services
 
+import io.webcontify.backend.collections.exceptions.UnprocessableContentException
 import io.webcontify.backend.collections.repositories.CollectionItemRepository
 import io.webcontify.backend.collections.utils.snakeToCamelCase
 import org.springframework.stereotype.Service
@@ -10,33 +11,41 @@ class CollectionItemService(
     val collectionItemRepository: CollectionItemRepository
 ) {
 
-  fun getById(collectionId: Int, identifierMap: Map<String, String?>): Map<String, Any> {
+  fun getById(collectionId: Int, identifierMap: Map<String, Any?>): Map<String, Any> {
     val collection = collectionService.getById(collectionId)
     return collectionItemRepository.getByIdFor(collection, identifierMap)
   }
 
-  fun getById(collectionId: Int, itemId: String): Map<String, Any> {
+  fun getById(collectionId: Int, itemId: Any): Map<String, Any> {
     val collection = collectionService.getById(collectionId)
+    if (collection.columns == null) {
+      throw UnprocessableContentException()
+    }
     val primaryKey = collection.columns.first { it.isPrimaryKey }
     return collectionItemRepository.getByIdFor(
         collection, mapOf(Pair(primaryKey.name.lowercase(), itemId)))
   }
 
-  fun deleteById(collectionId: Int, identifierMap: Map<String, String?>) {
+  fun deleteById(collectionId: Int, identifierMap: Map<String, Any?>) {
     val collection = collectionService.getById(collectionId)
-    collection.columns
-        .filter { it.isPrimaryKey }
-        .forEach {
-          if (!identifierMap.containsKey(it.name.lowercase().snakeToCamelCase())) {
-            throw RuntimeException("not all primary keys defined")
-          }
-        }
+    val primaryKeys = collection.columns?.filter { it.isPrimaryKey }
+    if (primaryKeys?.size != identifierMap.size) {
+      throw UnprocessableContentException()
+    }
+    primaryKeys.forEach {
+      if (!identifierMap.containsKey(it.name.snakeToCamelCase())) {
+        throw UnprocessableContentException()
+      }
+    }
     return collectionItemRepository.deleteById(collection, identifierMap)
   }
 
-  fun deleteById(collectionId: Int, itemId: String) {
+  fun deleteById(collectionId: Int, itemId: Any) {
     val collection = collectionService.getById(collectionId)
-    val primaryKey = collection.columns.first { it.isPrimaryKey } ?: throw RuntimeException()
+    if (collection.columns == null) {
+      throw UnprocessableContentException()
+    }
+    val primaryKey = collection.columns.first { it.isPrimaryKey }
     return collectionItemRepository.deleteById(
         collection, mapOf(Pair(primaryKey.name.lowercase(), itemId)))
   }
